requirements.txt
pyTelegramBotAPI==4.4.0
Flask==2.0.3
psycopg2-binary==2.9.3
